chore: every example is now following the same file structure
