+++ alternative-routes/queries.js
@@ -40,71 +40,6 @@ mutation newRoute{
   }
 `;
 
-export const queryRoute = qql`
-query getRoute($id: ID!) {
-  route(id: $id) {
-    alternatives {
-      polyline
-      charges
-      duration
-      distance
-      consumption
-      chargeTime
-      saving {
-        money
-        co2
-      }
-      legs{
-        distance
-        chargeTime
-        origin{
-          geometry{
-            type
-            coordinates
-          }
-        }
-        destination{
-          geometry
-          {
-            type
-            coordinates
-          }
-        }
-      }
-    }
-    route {
-      charges
-      saving {
-        money
-        co2
-      }
-      chargeTime
-      distance
-      duration
-      consumption
-      polyline
-      legs{
-        distance
-        chargeTime
-        origin{
-          geometry{
-            type
-            coordinates
-          }
-        }
-        destination{
-          geometry
-          {
-            type
-            coordinates
-          }
-        }
-      }
-    }
-    status
-  }
-}`;
-
 export const routeUpdate = qql`
 subscription routeUpdatedById($id: ID!){
   routeUpdatedById(id: $id) {
