+++ battery-capacity/client.js
@@ -4,7 +4,7 @@ import { createRoute, routeUpdate } from './queries';
 import { pipe, subscribe } from 'wonka';
 
 /**
- * For the purpose of this example we use urgl - lightweights GraphQL client.
+ * For the purpose of this example we use urql - lightweights GraphQL client.
  * To establish a connection with Chargetrip GraphQL API you need to have an API key.
  * The key in this example is a public one and gives access only to a part of our extensive database.
  * You need a registered `x-client-id` to access the full database.
@@ -41,9 +41,9 @@ const client = createClient({
  * @param { number } capacity - The current battery capacity
  * @param { function } callback - As soon as our asynchronous call is finished we do a callback to update our UI.
  */
-export const fetchRoute = (capacity, callback) => {
+export const getRoute = (capacity, callback) => {
   client
-    .mutation(createRoute(capacity))
+    .mutation(createRoute, { capacity: capacity })
     .toPromise()
     .then(response => {
       const routeId = response.data.newRoute;
@@ -52,6 +52,7 @@ export const fetchRoute = (capacity, callback) => {
         client.executeSubscription(createRequest(routeUpdate, { id: routeId })),
         subscribe(result => {
           const { status, route } = result.data?.routeUpdatedById;
+
           // You can keep listening to the route changes to update route information.
           // For this example we want to only draw the initial route.
           if (status === 'done' && route) {
