+++ battery-capacity/queries.js
@@ -12,40 +12,40 @@ import qql from 'graphql-tag';
  *   - one passenger in the car (drive alone)
  *   - an adjustable battery capacity
  */
-export const createRoute = capacity => `
-mutation newRoute{
-    newRoute(
-      input: {
-        ev: {
-          id: "5d161be5c9eef46132d9d20a"
-          battery: {
-            stateOfCharge: { value: ${capacity}, type: kwh }
-            capacity: { value: ${capacity}, type: kwh }
-          }
-          plugs: { chargingPower: 150, standard: TESLA_S }
-          adapters: [
-            { chargingPower: 150, standard: IEC_62196_T2_COMBO }
-            { chargingPower: 150, standard: CHADEMO }
-          ]
-          climate: true
-          numberOfPassengers: 1
+export const createRoute = qql`
+mutation newRoute($capacity: Float!){
+  newRoute(
+    input: {
+      ev: {
+        id: "5d161be5c9eef46132d9d20a"
+        battery: {
+          stateOfCharge: { value: $capacity, type: kwh }
+          capacity: { value: $capacity, type: kwh }
         }
-        routeRequest: {
-          origin: {
-            type: Feature
-            geometry: { type: Point, coordinates: [9.732625731357011, 52.3806314590276] }
-            properties: { name: "Hanover, Germany" }
+        plugs: { chargingPower: 150, standard: TESLA_S }
+        adapters: [
+          { chargingPower: 150, standard: IEC_62196_T2_COMBO }
+          { chargingPower: 150, standard: CHADEMO }
+        ]
+        climate: true
+        numberOfPassengers: 1
+      }
+      routeRequest: {
+        origin: {
+          type: Feature
+          geometry: { type: Point, coordinates: [9.732625731357011, 52.3806314590276] }
+          properties: { name: "Hanover, Germany" }
 
-          }
-          destination: {
-            type: Feature
-            geometry: { type: Point, coordinates: [9.922192327081783, 57.046057998779176] }
-            properties: { name: "Aalborg, Denmark" }
-          }
+        }
+        destination: {
+          type: Feature
+          geometry: { type: Point, coordinates: [9.922192327081783, 57.046057998779176] }
+          properties: { name: "Aalborg, Denmark" }
         }
       }
-    )
-  }
+    }
+  )
+}
 `;
 
 export const routeUpdate = qql`
