+++ elevation-plot/index.js
@@ -1,44 +1,23 @@
-import { drawRoute } from './map.js';
-import * as mapboxPolyline from '@mapbox/polyline';
 import { fetchRoute } from './client';
-import { renderRouteHeader, renderRouteDetails, loadGraph } from './elevationGraph';
+import { renderRouteHeader, renderRouteDetails, renderGraph, attachEventListeners } from './interface';
+import { drawRoutePolyline } from './map';
 
 /**
- * Draw a route on a map.
+ * This project shows you how to fetch a car list and render the car details
+ * The project structure contains;
  *
- * The route object contains `polyline` data -  the polyline encoded route (series of coordinates as a single string).
- * We need to decode this information first. We use Mapbox polyline tool (https://www.npmjs.com/package/@mapbox/polyline) for this.
- * As a result of decoding we get pairs [latitude, longitude].
- * To draw a route on a map we use Mapbox GL JS. This tool uses the format [longitude,latitude],
- * so we have to reverse every pair.
- *
- * @param { object } data - route specification.
- * @param { string } id - route ID.
+ *    - client.js - All networking requests
+ *    - interface.js - All interface rendering
+ *    - map.js - All map rendering (including routes and waypoints)
+ *    - queries.js - The GraphQL queries used in the networking requests
  */
+
 fetchRoute((routeId, routeData) => {
   drawRoutePolyline(routeId, routeData);
-  loadGraph(routeData);
+
+  renderRouteHeader(routeData);
+  renderRouteDetails(routeData.pathPlot[0]);
+  renderGraph(routeData);
 });
 
-const drawRoutePolyline = (id, data) => {
-  const decodedData = mapboxPolyline.decode(data?.polyline);
-  const reversed = decodedData.map(item => item.reverse());
-
-  drawRoute(id, reversed, data?.legs);
-  renderRouteHeader(data);
-  renderRouteDetails(data.pathPlot[0]);
-
-  const plotSliderInput = document.getElementById('slider-input');
-  const plotSliderOutput = document.getElementById('slider-output');
-
-  plotSliderInput.addEventListener('input', () => {
-    renderRouteDetails(data.pathPlot[plotSliderInput.value]);
-
-    const val = plotSliderInput.value;
-    const min = plotSliderInput.min ? plotSliderInput.min : 0;
-    const max = plotSliderInput.max ? plotSliderInput.max : 100;
-    const newVal = Number(((val - min) * 100) / (max - min));
-
-    plotSliderOutput.style.left = `calc(${newVal}% + (${11 - newVal * 0.25}px))`;
-  });
-};
+attachEventListeners();
