+++ tile-server/index.js
@@ -1,6 +1,9 @@
 import { displayMap } from './map';
 
 /**
- * Initializes the map with clusters and a polygon to render isolated Ecomovement data.
+ * This project shows you how to fetch a car list and render the car details
+ * The project structure contains;
+ *
+ *    - map.js - All map rendering (tile server)
  */
 displayMap();
